wip: Kizuna Samrt support
This is a BT2 serial rotation controller that doesn't
have a reverse mode, so it's mapped as a vibrator right now.

The 2 outputs are not indivually contrlled.

The device registers 2 RFCOMM ports, only the second seems
to allow connections.

Whilst I'm able to write to the device, I'm not getting any
response, so something is wrong.