feat: Allow for device reconnect
This change removes a race condition on insert into the connecting list
where 2 threads could check the DashSet then both insert. Since the insert
returns true on an actual insert, checking if the value is false should
make the previously 2 step operation atomic. The address is also now
removed from the set at 2 other points where we exit early.

It's also possible for the device to removed prior to it being added to
device_map, which would trigger the expect() in the device removed
callback. I hit that in debugging, where things are much slower than
in reality, but I believe the race condition is still possible.